requirements.txt
discord.py
SQLAlchemy
psycopg2-binary
pandas